refactor: renamed Header utility file
